% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{build_directory_url}
\alias{build_directory_url}
\title{Build school directory URL for a given year}
\usage{
build_directory_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
URL string
}
\description{
Constructs the URL for downloading the school directory file.
}
\keyword{internal}
