% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_multistats_url}
\alias{build_multistats_url}
\title{Build Multi Stats URL for a given year}
\usage{
build_multistats_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
URL string
}
\description{
Constructs the URL for downloading Multi Stats files based on year.
LDOE uses varying URL patterns across years.
}
\keyword{internal}
