% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if cached data exists for a year}
\usage{
cache_exists(end_year, cache_type = "tidy")
}
\arguments{
\item{end_year}{School year end}

\item{cache_type}{Type of cache ("tidy" or "wide")}
}
\value{
TRUE if cache exists
}
\description{
Check if cached data exists for a year
}
\keyword{internal}
