% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_status}
\alias{cache_status}
\title{View cache status}
\usage{
cache_status()
}
\value{
Data frame with cache status for each year
}
\description{
Shows which years have cached data.
}
\examples{
cache_status()
}
