% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_assessment.R
\name{calc_basic_above}
\alias{calc_basic_above}
\title{Calculate Basic and Above rates}
\usage{
calc_basic_above(df, ...)
}
\arguments{
\item{df}{Tidy assessment data frame}

\item{...}{Optional grouping variables}
}
\value{
Data frame with basic_above rates
}
\description{
Calculates the combined rate of students scoring Basic, Mastery, or Advanced.
This is sometimes used as a secondary proficiency indicator.
}
