% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{clean_assessment_names}
\alias{clean_assessment_names}
\title{Clean assessment column names}
\usage{
clean_assessment_names(x)
}
\arguments{
\item{x}{Vector of column names}
}
\value{
Cleaned column names
}
\description{
Clean assessment column names
}
\keyword{internal}
