% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enr_grade_aggs}
\alias{enr_grade_aggs}
\title{Create grade-level aggregations}
\usage{
enr_grade_aggs(df)
}
\arguments{
\item{df}{Data frame from tidy_enr}
}
\value{
Data frame with additional grade band rows
}
\description{
Creates summary rows by grade level bands (elementary, middle, high).
}
\keyword{internal}
