% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{find_directory_sheet}
\alias{find_directory_sheet}
\title{Find matching sheet name in directory Excel file}
\usage{
find_directory_sheet(sheets, patterns)
}
\arguments{
\item{sheets}{Vector of sheet names}

\item{patterns}{Vector of patterns to match}
}
\value{
Matching sheet name or NULL
}
\description{
Find matching sheet name in directory Excel file
}
\keyword{internal}
