% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_alternate_urls}
\alias{get_alternate_urls}
\title{Get alternate URL patterns for a year}
\usage{
get_alternate_urls(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Vector of alternate URLs
}
\description{
Returns a list of alternate URL patterns to try if primary URL fails.
}
\keyword{internal}
