% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Louisiana enrollment data}
\usage{
get_available_years()
}
\value{
A list with:
\item{min_year}{First available year (2019)}
\item{max_year}{Last available year (2024)}
\item{description}{Description of data availability}
}
\description{
Returns the range of years for which enrollment data is available
from the Louisiana Department of Education (LDOE).
}
\examples{
years <- get_available_years()
print(years$min_year)
print(years$max_year)
}
