% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get cache directory path}
\usage{
get_cache_dir()
}
\value{
Path to cache directory
}
\description{
Returns the path to the package cache directory.
}
\keyword{internal}
