% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_path}
\alias{get_cache_path}
\title{Get cache file path for a year}
\usage{
get_cache_path(end_year, cache_type = "tidy")
}
\arguments{
\item{end_year}{School year end}

\item{cache_type}{Type of cache ("tidy" or "wide")}
}
\value{
Path to cache file
}
\description{
Get cache file path for a year
}
\keyword{internal}
