% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_directory_available_years}
\alias{get_directory_available_years}
\title{Get available years for school directory data}
\usage{
get_directory_available_years()
}
\value{
A list with:
\item{min_year}{First available year (2025)}
\item{max_year}{Last available year (2025)}
\item{description}{Description of data availability}
}
\description{
Returns the range of years for which school directory data is available.
}
\examples{
years <- get_directory_available_years()
print(years$max_year)
}
