% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw school directory data from LDOE}
\usage{
get_raw_directory(end_year)
}
\arguments{
\item{end_year}{School year end (2024-25 = 2025)}
}
\value{
List with data frames for each sheet (public, charter, nonpublic)
}
\description{
Downloads the school directory Excel file from LDOE.
}
\keyword{internal}
