% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_directory_sheet}
\alias{process_directory_sheet}
\title{Process a single directory sheet}
\usage{
process_directory_sheet(df, end_year, is_charter = FALSE)
}
\arguments{
\item{df}{Data frame from a directory sheet}

\item{end_year}{School year end}

\item{is_charter}{Whether this is a charter school sheet}
}
\value{
Processed data frame
}
\description{
Process a single directory sheet
}
\keyword{internal}
