% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_cache}
\alias{read_cache}
\title{Read cached data for a year}
\usage{
read_cache(end_year, cache_type = "tidy")
}
\arguments{
\item{end_year}{School year end}

\item{cache_type}{Type of cache ("tidy" or "wide")}
}
\value{
Cached data frame
}
\description{
Read cached data for a year
}
\keyword{internal}
