% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{split_combined_data}
\alias{split_combined_data}
\title{Split combined data into site and LEA}
\usage{
split_combined_data(df, end_year)
}
\arguments{
\item{df}{Combined data frame}

\item{end_year}{School year end}
}
\value{
List with site and lea data frames
}
\description{
When data is in a single sheet, split based on content.
}
\keyword{internal}
