% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_la_grade}
\alias{standardize_la_grade}
\title{Standardize Louisiana grade levels}
\usage{
standardize_la_grade(x)
}
\arguments{
\item{x}{Vector of grade values}
}
\value{
Standardized grade levels
}
\description{
Standardize Louisiana grade levels
}
\keyword{internal}
