% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_la_subgroup}
\alias{standardize_la_subgroup}
\title{Standardize Louisiana subgroup names}
\usage{
standardize_la_subgroup(x)
}
\arguments{
\item{x}{Vector of subgroup names}
}
\value{
Standardized subgroup names
}
\description{
Standardize Louisiana subgroup names
}
\keyword{internal}
