% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_la_subject}
\alias{standardize_la_subject}
\title{Standardize Louisiana subject names}
\usage{
standardize_la_subject(x)
}
\arguments{
\item{x}{Vector of subject names}
}
\value{
Standardized subject names
}
\description{
Standardize Louisiana subject names
}
\keyword{internal}
