% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_cache}
\alias{write_cache}
\title{Write data to cache for a year}
\usage{
write_cache(df, end_year, cache_type = "tidy")
}
\arguments{
\item{df}{Data frame to cache}

\item{end_year}{School year end}

\item{cache_type}{Type of cache ("tidy" or "wide")}
}
\description{
Write data to cache for a year
}
\keyword{internal}
